Move `fetch-config` to `config`
